Change how instance definitions expand; add the Show typeclass
This changes how instances expand by breaking them into two steps, one
of which defines the instance, the other of which defines the instance
methods. This is mostly done to allow the `instance` form to expand to
a begin-for-syntax form that executes the instance registration as a
side-effect, which means the side-effect will be re-executed during
expansion of other importing modules.

One other potential advantage of this approach is that it defers the
expansion of method definitions somewhat, which might theoretically make
it easier to perfom partial expansion to assist with recursive
definitions and use-before-definition situations.