Change specification of main entrypoint to be a special form
Rather than specifying the entrypoint for a module by defining a
top-level binding with the name “main”, this introduces a new special
form called “main” that specifies an IO action to be executed when the
module is run as a program. This simplifies the way main is detected
considerably, and it makes it easier to implement the necessary
typechecking.