Change type parsing to treat constructors like values instead of macros
Until this change, type constructors have been parsed like macros. For
example, the type constructor Tuple was defined as a macro that expected
two type arguments and expanded into a fully-applied type. The
consequence of this is that all types had to appear fully-saturated in
order to result in a successful parse.

This change adjusts how types are parsed by changing the behavior of
application when in type-parsing mode. Types are now simply
variable-like transformers that expand into type constructors, and
application is what introduces τ:apps. This does not result in a
higher-kinded system on its own, but it is a first step in moving
towards it.