use a custom current-print
This new current-print is used both by the repl and by scribble. Instead
of calling printf, #%top-interaction now returns both the value and the
type, which is then printed by this new current-print.

This fixes the problem with eval:check printing the types twice due
to #%top-interaction being called on both the expression and the
expected result. #%top-interaction is still called twice, but it doesn't
print anything anymore, and current-print is only called once.

This also has a less pleasing consequence on the documentation: the
color of the ": type" line used to be different from the color of the
output value, but now they are the same color. This is unfortunate, but
I think this is for the best, because I think these different colors are
because scribble tries to distinguish the output produced as a
side-effect of evaluating an expression from the result of the
expression itself. So once the documentation contains examples which
print to the console, we will probably be happy that this output doesn't
have the same color as the ": type" line.

There is yet another side-effect on the documentation, namely that the
output values are now printed in a different style, for example
"#<true>" instead of "(true)". I don't know which of those two styles is
best (probably neither), but the repl's output style used to be and
still is "#<true>", so now the documentation matches reality.