Add letrec
This ended up requiring a larger restructuring of the typechecking code
than I had hoped. Essentially, typechecking mutually recursive bindings
requires expanding them all at once, so they can share the same set of
bindings/assumptions, and some of them might need to be inferred while
others need to be checked.

To accommodate that, this commit merges and generalizes τ⇒/λ! and τ⇐/λ!
into a new function, τs⇔/λ!. It has the power to infer and check many
expressions at once under the same set of bindings, necessary for letrec
and probably little else. The other functions are now defined in terms
of τs⇔/λ!.

related to #28