typechecker: Don’t track ctx:assump in the type context
From the beginning of this code’s existence, this was a remnant of the
Haskell implementation, and it shows — removing this was remarkably
straightforward. There is no need to track the types of identifiers in
the type context, since types are attached to identifiers with syntax
properties. The context is essentially implicit and maintained by the
existing Racket model of lexical context, so assumptions are redundant.