Add todo!
A todo! form marks a region of a program that is not yet written,
in which the programmer can write whatever they want. The intention
behind todo! is that it be used for fancy interactive editing features
Haskell has a similar feature, which it calls "typed holes", which
don't allow any contents. These ones are more similar to "sheds" found
in Epigram or Agda.

At present, a todo! form expands to a program that throws an error that
lists the expected type. However, it also adds a syntax property that
tools such as the todo list can use to provide a list of todos and
their goal types.

Future enhancements can include diffing the type synthesized for the
todo! form’s contents against the type expected for the todo! form
itself, and adding the typing context to the message and/or goal.