Add typeclass deriving for Eq, and add support for context reduction
Context reduction was necessary for a clean implementation of Eq
deriving, since it revealed a tendency to produce cyclic constraints,
such as in the derived Eq instance on lists:

  (instance (forall [a] (Eq a) (Eq (List a)) => (Eq (List a))) ....)

The redundant (Eq (List a)) constraint in the instance head would cause
the typechecker to go into an infinite loop. We could just make instance
prune such constraints, but I took this as an opportunity to solve a
the more general problem of context reduction to make instance a little
less special.