Change #%type:forall to make it more like a real binding form
Semantically, #%type:forall is like a heavily restricted type-level
lambda, and it functions like a binding form in the same way that lambda
does. This commit adjusts the implementation of #%type:forall so it
actually does function like a binding form.

The most significant effect of this is that bound type variables are no
longer represented separately in the type language, so #%type:bound-var
has been removed entirely. Instead, bound type variables are merely
identifiers left behind after type expansion.