Rewrite elaboration to use a new, multi-pass expansion strategy
This technique generalizes the existing approach to elaboration to
elaborate programs in three passes (the second of which actually runs an
arbitrary number of times until elaboration reaches a fixpoint). This
allows macros to cooperate with the elaborator, and it allows the
implementation of elaborating macros that can provide additional
information to the typechecker (such as, for example, functional
dependencies).